Backend, refactor: 1) Move a project existence check to Project model; refactor&fix: 2) Change the keys format for caching (after switching to collaboration, the old key format is not suitable for tasks due to different user roles).
Frontend, feat: 1) Add a user check for sufficient rights to perform an action on the task page (including the task creation/modification modals).
The main problem at the moment is the inability to remove the create button for a user who isn't authorized to create a new task.
Now it's simply not clickable.